대댓글 작성, 조회, 수정, 삭제 구현

+++ cafe/src/main/kotlin/com/example/cafe/comment/service/CommentServiceImpl.kt
@@ -3,6 +3,8 @@ package com.example.cafe.comment.service
 import com.example.cafe.article.repository.ArticleRepository
 import com.example.cafe.comment.repository.CommentEntity
 import com.example.cafe.comment.repository.CommentRepository
+import com.example.cafe.comment.repository.RecommentEntity
+import com.example.cafe.comment.repository.RecommentRepository
 import com.example.cafe.user.repository.UserRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -11,13 +13,10 @@ import java.time.LocalDateTime
 @Service
 class CommentServiceImpl (
     private val commentRepository: CommentRepository,
+    private val recommentRepository: RecommentRepository,
     private val userRepository: UserRepository,
     private val articleRepository: ArticleRepository,
 ) : CommentService {
-    override fun getComment(id: Long): Comment {
-        TODO()
-    }
-
     override fun getComments(articleId: Long): List<Comment> {
         val article = articleRepository.findByIdOrNull(articleId) ?: throw CommentArticleNotFoundException()
         val comments = article.comments.map{ commentEntity ->
@@ -39,12 +38,17 @@ class CommentServiceImpl (
         return comments
     }
 
-    override fun getRecomment(id: Long): Recomment {
-        TODO()
-    }
-
     override fun getRecomments(commentId: Long): List<Recomment> {
-        TODO()
+        val comment = commentRepository.findByIdOrNull(commentId) ?: throw CommentNotFoundException()
+        val recomments = comment.recomments.map{ recommentEntity ->
+            Recomment(
+                id = recommentEntity.id,
+                content = recommentEntity.content,
+                lastModified = recommentEntity.lastModified,
+                nickname = recommentEntity.user.nickname,
+            )
+        }
+        return recomments
     }
 
     override fun createComment(userId: String, articleId: Long, content: String, at: LocalDateTime): Comment {
@@ -67,7 +71,22 @@ class CommentServiceImpl (
     }
 
     override fun createRecomment(userId: String, commentId: Long, content: String, at: LocalDateTime) : Recomment {
-        TODO()
+        val user = userRepository.findByUserId(userId) ?: throw CommentUserNotFoundException()
+        val comment = commentRepository.findByIdOrNull(commentId) ?: throw CommentNotFoundException()
+        val recomment = recommentRepository.save(
+            RecommentEntity(
+                content = content,
+                lastModified = at,
+                user = user,
+                comment = comment,
+            )
+        )
+        return Recomment(
+            id = recomment.id,
+            content = recomment.content,
+            lastModified = recomment.lastModified,
+            nickname = recomment.user.nickname,
+        )
     }
 
     override fun updateComment(id: Long, userId: String, content: String, at: LocalDateTime) : Comment {
@@ -94,15 +113,33 @@ class CommentServiceImpl (
     }
 
     override fun updateRecomment(id: Long, userId: String, content: String, at: LocalDateTime) : Recomment {
-        TODO()
+        val recomment = recommentRepository.findByIdOrNull(id) ?: throw RecommentNotFoundException()
+        if (recomment.user.userId != userId) {
+            throw InvalidCommentUserException()
+        }
+        recomment.content = content
+        recomment.lastModified = at
+        return Recomment(
+            id = recomment.id,
+            content = recomment.content,
+            lastModified = recomment.lastModified,
+            nickname = recomment.user.nickname,
+        )
     }
 
     override fun deleteComment(id: Long, userId: String) {
         val comment = commentRepository.findByIdOrNull(id) ?: throw CommentNotFoundException()
+        if (comment.user.userId != userId) {
+            throw InvalidCommentUserException()
+        }
         commentRepository.delete(comment)
     }
 
     override fun deleteRecomment(id: Long, userId: String) {
-        TODO()
+        val recomment = recommentRepository.findByIdOrNull(id) ?: throw RecommentNotFoundException()
+        if (recomment.user.userId != userId) {
+            throw InvalidCommentUserException()
+        }
+        recommentRepository.delete(recomment)
     }
 }
